refactor(hume-rope): namespace lines/cursor modules, unify tab_width on u8
Drop the crate-root pub use flattening of `lines`/`cursor` in favor of
`pub mod` — matches the existing `grapheme`/`width`/`position_encoding`
convention and removes hume-editing's strip_line_break/CharCursor
re-exports. Unify width.rs's tab_width parameter on u8 (grapheme.rs and
lines.rs already used u8; width.rs alone used usize), dropping the
scattered `as usize` casts at every call site. Narrow grapheme_count to
pub(crate) — it has no external callers, only grapheme_col_in_line uses it.

First of 5 commits consolidating column computation/conversion code into
hume-rope.

+++ hume-engine/src/render.rs
@@ -212,7 +212,7 @@ fn compose_gutter(
                 let w = hume_rope::width::grapheme_width(
                     g,
                     text_width as usize,
-                    compose_ctx.tab_width as usize,
+                    compose_ctx.tab_width,
                 ) as u16;
                 if text_width + w > usable_per_cell {
                     truncated_len = byte_idx;
